Migrate columns to separate store
